SCB-1113 add AlphaCluster and client properties.
