+++ omega/omega-configuration/src/main/java/org/apache/servicecomb/pack/omega/configuration/AlphaClusterProperties.java
@@ -1,0 +1,44 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.servicecomb.pack.omega.configuration;
+
+import java.util.Collections;
+import java.util.List;
+
+public class AlphaClusterProperties {
+
+  private List<String> addresses = Collections.singletonList("localhost:8080");
+
+  private AlphaSSLProperties ssl;
+
+  public List<String> getAddresses() {
+    return addresses;
+  }
+
+  public void setAddresses(List<String> addresses) {
+    this.addresses = addresses;
+  }
+
+  public AlphaSSLProperties getSsl() {
+    return ssl;
+  }
+
+  public void setSsl(AlphaSSLProperties ssl) {
+    this.ssl = ssl;
+  }
+}
