+++ omega/omega-spring-starter/pom.xml
@@ -32,11 +32,20 @@
   <dependencies>
     <dependency>
       <groupId>org.apache.servicecomb.pack</groupId>
+      <artifactId>omega-configuration</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.servicecomb.pack</groupId>
       <artifactId>omega-spring-tx</artifactId>
     </dependency>
     <dependency>
       <groupId>org.apache.servicecomb.pack</groupId>
       <artifactId>omega-format</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-configuration-processor</artifactId>
+      <optional>true</optional>
     </dependency>
     <dependency>
       <groupId>io.grpc</groupId>
