+++ omega/pom.xml
@@ -72,6 +72,11 @@
         <artifactId>omega-transaction</artifactId>
         <version>0.4.0-SNAPSHOT</version>
       </dependency>
+      <dependency>
+        <groupId>org.apache.servicecomb.pack</groupId>
+        <artifactId>omega-configuration</artifactId>
+        <version>0.4.0-SNAPSHOT</version>
+      </dependency>
     </dependencies>
   </dependencyManagement>
